port some functions to common/Components.kt
